Remodelled Game
Game has been rebuilt to be more in-line with what is now expected of
it. It has been modified to include more information, it also been
designed to include changes that are relevant to the graphics, e.g.
the changes that have been made to the board, player movements.

Java Doc Not complete.