Updates to File Readers/Writers
File Writer is now done. Board and Game and Game State have had some
updates done to reduce the amount of data that is stored (removed
unnecessary data items).

ActionTilePlaceable has been updated to ensure that a tile in action can
have the time set.

FileReader has been updated to match it's counter part in file writer,
but is not yet complete.